아이템 추가 및 강화기능 구현

+++ Assets/Script/Game/Unit/UnitPrefap.cs
@@ -6,11 +6,11 @@ using UnityEngine;
 public class UnitPrefap : ScriptableObject
 {
     public Sprite portrait;
-    public double hp;
-    public double attackDamage;
-    public double criticalRate;
-    public double criticalDamage;
-    public double defence;
+    public float hp;
+    public float attackDamage;
+    public float criticalRate;
+    public float criticalDamage;
+    public float defence;
     public string explain;
     public Item[] dropItems;
 }
